Получение информации о версиях и работа с файлом лога обновлений выделеены в отдельный класс
